Remove utils file
