+++ mftool.py
@@ -19,11 +19,11 @@
 """
 # -*- coding: UTF-8 -*-
 import os
-from utils import get_friday, get_today, is_holiday, render_response
 import requests
 import json
 from bs4 import BeautifulSoup
 import datetime
+from datetime import date,timedelta
 
 class Mftool():
     """
@@ -411,3 +411,23 @@ class Mftool():
                 scheme_category[scheme['categoryName']].append(scheme_info.copy())
                 scheme_info = None
         return render_response(scheme_category, as_json)
+
+def is_holiday():
+    if date.today().strftime("%a") in ['Sat', 'Sun', 'Mon']:
+        return True
+    else:
+        return False
+
+def get_friday():
+    days = {'Sat': 1, 'Sun': 2, 'Mon': 3}
+    diff = int(days[date.today().strftime("%a")])
+    return (date.today() - timedelta(days=diff)).strftime("%d-%b-%Y")
+
+def get_today():
+    return (date.today() - timedelta(days=1)).strftime("%d-%b-%Y")
+
+def render_response(data, as_json=False):
+    if as_json is True:
+        return json.dumps(data)
+    else:
+        return data
